Split features into components
